update router interface
