+++ core/server.go
@@ -163,9 +163,7 @@ func (s *Server) handleFrameConn(fconn frame.Conn, logger *slog.Logger) {
 	s.connHandler(conn) // s.handleConnRoute(conn) with middlewares
 
 	if conn.ClientType() == ClientTypeStreamFunction {
-		_ = s.router.Remove(&router.RouteParams{
-			ID: conn.ID(),
-		})
+		s.router.Remove(conn.ID())
 	}
 	_ = s.connector.Remove(conn.ID())
 }
@@ -260,12 +258,7 @@ func (s *Server) addSfnToRoute(hf *frame.HandshakeFrame, md metadata.M) error {
 	if hf.ClientType != byte(ClientTypeStreamFunction) {
 		return nil
 	}
-	return s.router.Add(&router.RouteParams{
-		Name:            hf.Name,
-		ID:              hf.ID,
-		Metadata:        md,
-		ObserveDataTags: hf.ObserveDataTags,
-	})
+	return s.router.Add(hf.ID, hf.ObserveDataTags, md)
 }
 
 func negotiateVersion(cVersion, sVersion string) error {
@@ -321,9 +314,7 @@ func (s *Server) routingDataFrame(c *Context) error {
 	dataFrame.Metadata = mdBytes
 
 	// find stream function ids from the router.
-	connIDs := s.router.Get(&router.RouteParams{
-		ObserveDataTags: []uint32{dataFrame.Tag},
-	})
+	connIDs := s.router.Route(dataFrame.Tag, md)
 	if len(connIDs) == 0 {
 		c.Logger.Info("no observed", "tag", dataFrame.Tag, "data_length", data_length)
 	}
